% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuickMap.R
\name{QuickMap}
\alias{QuickMap}
\title{QuickMap
plot a quick leaflet map of R}
\usage{
QuickMap(sf_data, color_str)
}
\arguments{
\item{sf_data}{sf}

\item{color_str}{str}
}
\value{
leaflet
}
\description{
QuickMap
plot a quick leaflet map of R
}
\examples{
QuickMap(spData::world, "red")

}
